PM: winning ticket validation WIP
We had to upgrade to a newer Ganache version to receive revert error
messages in our tests.
We had to upgrade to a new Truffle version to enjoy the feature of
calling functions with struct objects. As part of this upgrade we also
had to change more BigNumber references to BN.

Signature verification still in progress.