PM: TicketBroker inheritance refactor
So far we've used two different approaches for reusing code from
`TicketBroker` in its subtypes: one where only subtypes implement
functions and use modifiers to reuse `TicketBroker`'s flow of
requirements and accounting, and another approach where `TicketBroker`
implements a public/external function that contains the entire flow,
and calls an abstract method that each subtype implements according to
their needs.

The first approach (using modifiers) resulted in code duplication, i.e.
having to use the exact same modifiers in more than one contract, which
also resulted in duplicate test code.

To get out of this code duplication we're refactoring to the second
approach across the board.

The main cost is the complication introduced in the ERC20 implementation,
but since it's just an example and not a priority, it seems like the
better choice, for the benefit of simplifying the more mainstream
usecase with ETH.