PM: senders can revoke signers
We decided to create a separate param for the signer revocation delay
to decouple this concern from the `unlockPeriod`.
The reason we need a delayed revocation in the first place, is to allow
receivers of tickets to become aware of a pending revocation and make
sure they cash in any tickets signed by the soon-to-be-revoked signer
before those tickets become invalid.